refactor: move back the main at the root of the package
