+++ cli/chi_util.go
@@ -1,4 +1,4 @@
-package main
+package cli
 
 import (
 	"net/http"
