+++ cli/cmd_db.go
@@ -1,4 +1,4 @@
-package main
+package cli
 
 import (
 	"encoding/json"
