+++ cli/cmd_pull.go
@@ -1,4 +1,4 @@
-package main
+package cli
 
 import (
 	"encoding/json"
